web/requirements.txt
django==1.10.1
msgpack-rpc-python
beanstalkc
msgpack
redis
suds